+++ pdi/include/pdi/context.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2015-2024 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2015-2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
  * Copyright (C) 2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
@@ -37,6 +37,7 @@
 #include <pdi/datatype_template.h>
 #include <pdi/logger.h>
 #include <pdi/ref_any.h>
+#include <pdi/timer.h>
 
 namespace PDI {
 
@@ -113,6 +114,11 @@ public:
 	 * \return logger
 	 */
 	virtual Logger& logger() = 0;
+
+	/** Timer getter
+	 * \return timer
+	 */
+	virtual Timer& timer() = 0;
 
 	/** Callbacks of the context
 	 * \return context callbacks