+++ pdi/src/pdi.cxx
@@ -42,6 +42,7 @@
 #include "pdi/paraconf_wrapper.h"
 #include "pdi/plugin.h"
 #include "pdi/ref_any.h"
+#include "pdi/timer.h"
 
 #include "global_context.h"
 
@@ -170,6 +171,7 @@ try {
 	g_transaction.clear();
 	g_transaction_data.clear();
 	Global_context::init(conf);
+	Global_context::context().timer().startTimer("PDI");
 	return PDI_OK;
 } catch (const Error& e) {
 	return g_error_context.return_err(e);
@@ -184,6 +186,8 @@ try {
 	Paraconf_wrapper fw;
 	g_transaction.clear();
 	g_transaction_data.clear();
+	Global_context::context().timer().stopTimer("PDI");
+	Global_context::context().timer().printReport();
 	Global_context::finalize();
 	return PDI_OK;
 } catch (const Error& e) {