+++ plugins/decl_hdf5/decl_hdf5.cxx
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2015-2019 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2015-2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
  * Copyright (C) 2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
@@ -114,18 +114,22 @@ public:
 
 	void data(const std::string& name, Ref ref)
 	{
+		context().timer().startTimer(pretty_name());
 		Hdf5_error_handler _;
 		for (auto&& op: m_data[name]) {
 			op.execute(context());
 		}
+		context().timer().stopTimer(pretty_name());
 	}
 
 	void event(const std::string& event)
 	{
+		context().timer().startTimer(pretty_name());
 		Hdf5_error_handler _;
 		for (auto&& op: m_events[event]) {
 			op.execute(context());
 		}
+		context().timer().stopTimer(pretty_name());
 	}
 
 	/** Pretty name for the plugin that will be shown in the logger