+++ plugins/serialize/serialize.cxx
@@ -272,6 +272,7 @@ struct serialize_plugin: PDI::Plugin {
 	 */
 	void share_serialized(const std::string& desc_name, PDI::Ref ref)
 	{
+		context().timer().startTimer(pretty_name());
 		std::string serialized_name = m_desc_to_serialize[desc_name];
 		context().logger().debug("Serializing `{}` as `{}`", desc_name, serialized_name);
 		PDI::Datatype_sptr serialized_type = serialize_type(ref.type());
@@ -333,6 +334,7 @@ struct serialize_plugin: PDI::Plugin {
 			);
 			m_serialized_remove_callback.emplace_back(serialized_name, remove_callback, PDI_IN);
 		}
+		context().timer().stopTimer(pretty_name());
 	}
 
 	void release_serialized(const std::string& desc_name, PDI::Ref ref)